feat: Mejorar consistencia visual del sidebar en EventsManagement
- Arreglar espaciado de elementos del menú de py-2 a py-3 para consistencia
- Aumentar tamaño del icono del header de w-6 h-6 a w-8 h-8 para uniformidad
- Corregir espaciado del botón "Cerrar Sesión" de py-2 a py-3
- Alinear estilos con Configuration.tsx para mantener coherencia visual
- Mejorar experiencia de usuario con espaciado uniforme en todos los sidebars

+++ backend/authentication/repositories/stats.repository.ts
@@ -165,9 +165,9 @@ class StatsRepositoryImpl implements StatsRepository {
       )
       SELECT 
         TO_CHAR(ms.month_start, 'Month') as month,
-        EXTRACT(YEAR FROM ms.month_start) as year,
-        COALESCE(md.participants, 0) as participants,
-        COALESCE(md.events, 0) as events
+        EXTRACT(YEAR FROM ms.month_start)::integer as year,
+        COALESCE(md.participants, 0)::integer as participants,
+        COALESCE(md.events, 0)::integer as events
       FROM month_series ms
       LEFT JOIN monthly_data md ON ms.month_start = md.month
       ORDER BY ms.month_start
